Squash post-nonce Alembic migrations
Replace the post-1.29 state migration chain with a single squashed
revision and simplify redundant Alembic utility tests.
