--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_04_24_add_task_table_to_core_state.py
@@ -1,60 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Add task table to core state.
-
-Revision ID: 02aad1cc2a99
-Revises: b277e6f3656c
-Create Date: 2026-04-24 13:31:18.645691
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "02aad1cc2a99"
-down_revision: str | Sequence[str] | None = "b277e6f3656c"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "task",
-        sa.Column("task_id", sa.Integer(), nullable=False),
-        sa.Column("type", sa.String(), nullable=False),
-        sa.Column("run_id", sa.Integer(), nullable=False),
-        sa.Column("fab_hash", sa.String(), nullable=True),
-        sa.Column("model_ref", sa.String(), nullable=True),
-        sa.Column("connector_ref", sa.String(), nullable=True),
-        sa.Column("token", sa.String(), nullable=True),
-        sa.Column("pending_at", sa.String(), nullable=False),
-        sa.Column("starting_at", sa.String(), nullable=True),
-        sa.Column("running_at", sa.String(), nullable=True),
-        sa.Column("finished_at", sa.String(), nullable=True),
-        sa.UniqueConstraint("task_id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("task")
-    # ### end Alembic commands ###
