--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_04_24_widen_integer_columns_to_biginteger.py
@@ -1,333 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Widen Integer columns to BigInteger.
-
-Revision ID: dee9b802b5c9
-Revises: fb960f5973fc
-Create Date: 2026-04-24 08:49:50.976650
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "dee9b802b5c9"
-down_revision: str | Sequence[str] | None = "02aad1cc2a99"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-_RUN_ID_FOREIGN_KEYS = (
-    ("context", "context_run_id_fkey"),
-    ("logs", "logs_run_id_fkey"),
-    ("message_ins", "message_ins_run_id_fkey"),
-    ("message_res", "message_res_run_id_fkey"),
-)
-
-
-def _is_postgresql() -> bool:
-    """Return True if the migration is running against PostgreSQL."""
-    return op.get_bind().dialect.name == "postgresql"
-
-
-def _drop_run_id_foreign_keys() -> None:
-    """Drop run_id foreign keys before changing column types on PostgreSQL."""
-    if not _is_postgresql():
-        return
-    # These child tables reference run.run_id. PostgreSQL checks FK type
-    # compatibility during ALTER COLUMN, so widening one side at a time can
-    # create an intermediate INTEGER/BIGINT mismatch and fail the migration.
-    for table_name, constraint_name in _RUN_ID_FOREIGN_KEYS:
-        op.drop_constraint(constraint_name, table_name, type_="foreignkey")
-
-
-def _create_run_id_foreign_keys() -> None:
-    """Recreate run_id foreign keys after changing column types on PostgreSQL."""
-    if not _is_postgresql():
-        return
-    for table_name, constraint_name in _RUN_ID_FOREIGN_KEYS:
-        op.create_foreign_key(
-            constraint_name,
-            table_name,
-            "run",
-            ["run_id"],
-            ["run_id"],
-        )
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    _drop_run_id_foreign_keys()
-
-    with op.batch_alter_table("context", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("logs", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("message_ins", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "src_node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "dst_node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("message_res", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "src_node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "dst_node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("node", schema=None) as batch_op:
-        batch_op.alter_column(
-            "node_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "bytes_sent",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-            existing_server_default=sa.text("'0'"),
-        )
-        batch_op.alter_column(
-            "bytes_recv",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-            existing_server_default=sa.text("'0'"),
-        )
-
-    with op.batch_alter_table("run_objects", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=False,
-        )
-
-    with op.batch_alter_table("task", schema=None) as batch_op:
-        batch_op.alter_column(
-            "task_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=False,
-        )
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=False,
-        )
-
-    with op.batch_alter_table("token_store", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.INTEGER(),
-            type_=sa.BigInteger(),
-            existing_nullable=True,
-        )
-
-    _create_run_id_foreign_keys()
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    _drop_run_id_foreign_keys()
-
-    with op.batch_alter_table("token_store", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("task", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=False,
-        )
-        batch_op.alter_column(
-            "task_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=False,
-        )
-
-    with op.batch_alter_table("run_objects", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=False,
-        )
-
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.alter_column(
-            "bytes_recv",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-            existing_server_default=sa.text("'0'"),
-        )
-        batch_op.alter_column(
-            "bytes_sent",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-            existing_server_default=sa.text("'0'"),
-        )
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("node", schema=None) as batch_op:
-        batch_op.alter_column(
-            "node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("message_res", schema=None) as batch_op:
-        batch_op.alter_column(
-            "dst_node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "src_node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("message_ins", schema=None) as batch_op:
-        batch_op.alter_column(
-            "dst_node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "src_node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("logs", schema=None) as batch_op:
-        batch_op.alter_column(
-            "node_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    with op.batch_alter_table("context", schema=None) as batch_op:
-        batch_op.alter_column(
-            "run_id",
-            existing_type=sa.BigInteger(),
-            type_=sa.INTEGER(),
-            existing_nullable=True,
-        )
-
-    _create_run_id_foreign_keys()
-
-    # ### end Alembic commands ###
