--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_04_28_add_active_until_to_task_table.py
@@ -1,50 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Add active_until to task table.
-
-Revision ID: c7eb009ce75c
-Revises: 8253e456d570
-Create Date: 2026-04-28 19:24:40.811386
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "c7eb009ce75c"
-down_revision: str | Sequence[str] | None = "8253e456d570"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("task", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("active_until", sa.BigInteger(), nullable=True))
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("task", schema=None) as batch_op:
-        batch_op.drop_column("active_until")
-
-    # ### end Alembic commands ###
