--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_04_30_make_run_substatus_and_details_not_nullable.py
@@ -1,69 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Make run sub_status and details not nullable.
-
-Revision ID: e566a1e755c5
-Revises: c7eb009ce75c
-Create Date: 2026-04-30 09:43:42.202183
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "e566a1e755c5"
-down_revision: str | Sequence[str] | None = "c7eb009ce75c"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # Backfill existing NULLs before enforcing NOT NULL.
-    op.execute("UPDATE run SET sub_status = '' WHERE sub_status IS NULL")
-    op.execute("UPDATE run SET details = '' WHERE details IS NULL")
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.alter_column(
-            "sub_status",
-            existing_type=sa.VARCHAR(),
-            server_default=sa.text("''"),
-            nullable=False,
-        )
-        batch_op.alter_column(
-            "details",
-            existing_type=sa.VARCHAR(),
-            server_default=sa.text("''"),
-            nullable=False,
-        )
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.alter_column(
-            "details", existing_type=sa.VARCHAR(), server_default=None, nullable=True
-        )
-        batch_op.alter_column(
-            "sub_status", existing_type=sa.VARCHAR(), server_default=None, nullable=True
-        )
-
-    # ### end Alembic commands ###
