--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_08_add_logs_table_to_corestate.py
@@ -1,63 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Add logs table to corestate.
-
-Revision ID: 11d1bdac4c8e
-Revises: e566a1e755c5
-Create Date: 2026-05-08 21:17:30.106881
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "11d1bdac4c8e"
-down_revision: str | Sequence[str] | None = "e566a1e755c5"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "task_logs",
-        sa.Column("timestamp", sa.Float(), nullable=False),
-        sa.Column("task_id", sa.BigInteger(), nullable=False),
-        sa.Column("log", sa.String(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["task_id"],
-            ["task.task_id"],
-        ),
-    )
-    with op.batch_alter_table("task_logs", schema=None) as batch_op:
-        batch_op.create_index(
-            "idx_task_logs_task_id_timestamp", ["task_id", "timestamp"], unique=False
-        )
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("task_logs", schema=None) as batch_op:
-        batch_op.drop_index("idx_task_logs_task_id_timestamp")
-
-    op.drop_table("task_logs")
-    # ### end Alembic commands ###
