--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_08_drop_token_store_table.py
@@ -1,53 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Drop token_store table.
-
-Revision ID: dd247761db4c
-Revises: 8b40f767ddcb
-Create Date: 2026-05-08 11:10:09.790851
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "dd247761db4c"
-down_revision: str | Sequence[str] | None = "8b40f767ddcb"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("token_store")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "token_store",
-        sa.Column("run_id", sa.BIGINT(), nullable=True),
-        sa.Column("token", sa.VARCHAR(), nullable=False),
-        sa.Column("active_until", sa.FLOAT(), nullable=True),
-        sa.PrimaryKeyConstraint("run_id"),
-        sa.UniqueConstraint("token"),
-    )
-    # ### end Alembic commands ###
