--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_13_add_task_message_table.py
@@ -1,79 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Add task message table.
-
-Revision ID: dd6b78efd1f6
-Revises: dd247761db4c
-Create Date: 2026-05-13 11:19:23.959850
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "dd6b78efd1f6"
-down_revision: str | Sequence[str] | None = "dd247761db4c"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "task_message",
-        sa.Column("message_id", sa.String(), nullable=False),
-        sa.Column("run_id", sa.BigInteger(), nullable=False),
-        sa.Column("src_task_id", sa.BigInteger(), nullable=False),
-        sa.Column("dst_task_id", sa.BigInteger(), nullable=False),
-        sa.Column("reply_to_message_id", sa.String(), nullable=True),
-        sa.Column("created_at", sa.Float(), nullable=False),
-        sa.Column("ttl", sa.Float(), nullable=False),
-        sa.Column("message_type", sa.String(), nullable=False),
-        sa.Column("content", sa.LargeBinary(), nullable=True),
-        sa.Column("error", sa.LargeBinary(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["dst_task_id"],
-            ["task.task_id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["src_task_id"],
-            ["task.task_id"],
-        ),
-        sa.PrimaryKeyConstraint("message_id"),
-    )
-    with op.batch_alter_table("task_message", schema=None) as batch_op:
-        batch_op.create_index(
-            "idx_task_message_dst_task_id_created_at",
-            ["dst_task_id", "created_at"],
-            unique=False,
-        )
-        batch_op.create_index("idx_task_message_run_id", ["run_id"], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("task_message", schema=None) as batch_op:
-        batch_op.drop_index("idx_task_message_run_id")
-        batch_op.drop_index("idx_task_message_dst_task_id_created_at")
-
-    op.drop_table("task_message")
-    # ### end Alembic commands ###
