--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_13_migrate_logs_to_task_logs.py
@@ -1,70 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Migrate logs to task logs.
-
-Revision ID: aac61834ee69
-Revises: dd6b78efd1f6
-Create Date: 2026-05-13 11:43:20.602486
-"""
-from collections.abc import Sequence
-
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "aac61834ee69"
-down_revision: str | Sequence[str] | None = "dd6b78efd1f6"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.execute(
-        """
-        INSERT INTO task_logs (timestamp, task_id, log)
-        SELECT logs.timestamp, run.primary_task_id, logs.log
-        FROM logs
-        JOIN run ON run.run_id = logs.run_id
-        JOIN task ON task.task_id = run.primary_task_id
-        WHERE logs.timestamp IS NOT NULL
-          AND logs.log IS NOT NULL
-        """
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.execute(
-        """
-        INSERT INTO logs (timestamp, run_id, node_id, log)
-        SELECT task_logs.timestamp, task.run_id, 0, task_logs.log
-        FROM task_logs
-        JOIN task ON task.task_id = task_logs.task_id
-        WHERE NOT EXISTS (
-            SELECT 1
-            FROM logs
-            WHERE logs.timestamp = task_logs.timestamp
-              AND logs.run_id = task.run_id
-              AND logs.node_id = 0
-        )
-        """
-    )
-    op.execute("DELETE FROM task_logs")
-    # ### end Alembic commands ###
