--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_14_use_timestamp_columns_for_task.py
@@ -1,137 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Use timestamp columns for task timestamps.
-
-Revision ID: e937d3528d23
-Revises: aac61834ee69
-Create Date: 2026-05-14 15:15:35.278372
-"""
-from collections.abc import Sequence
-from typing import Any
-
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.sql.type_api import TypeEngine
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "e937d3528d23"
-down_revision: str | Sequence[str] | None = "aac61834ee69"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-TASK_TIMESTAMP_COLUMNS = ("pending_at", "starting_at", "running_at", "finished_at")
-# Each tuple contains: (column name, old type, new type, nullable in new schema)
-ColumnSpec = tuple[str, TypeEngine[Any], TypeEngine[Any], bool]
-TASK_TIMESTAMP_COLUMN_SPECS: tuple[ColumnSpec, ...] = (
-    ("active_until", sa.BIGINT(), sa.TIMESTAMP(timezone=True), True),
-    ("pending_at", sa.VARCHAR(), sa.TIMESTAMP(timezone=True), False),
-    ("starting_at", sa.VARCHAR(), sa.TIMESTAMP(timezone=True), True),
-    ("running_at", sa.VARCHAR(), sa.TIMESTAMP(timezone=True), True),
-    ("finished_at", sa.VARCHAR(), sa.TIMESTAMP(timezone=True), True),
-)
-REP_COL_PREFIX = "_new_"
-
-
-def _rep_col(column: str) -> str:
-    """Return the temporary replacement column name."""
-    return f"{REP_COL_PREFIX}{column}"
-
-
-def _add_replacement_columns(is_upgrade: bool) -> None:
-    """Add replacement columns with the appropriate timestamp types."""
-    for column, old_type, new_type, _ in TASK_TIMESTAMP_COLUMN_SPECS:
-        column_type = new_type if is_upgrade else old_type
-        op.add_column("task", sa.Column(_rep_col(column), column_type, nullable=True))
-
-
-def _replace_task_timestamp_columns(is_upgrade: bool) -> None:
-    """Swap old task timestamp columns with their replacement columns."""
-    # Some databases cannot directly convert integer columns to TIMESTAMP, so use
-    # replacement columns instead of in-place type conversion.
-    with op.batch_alter_table("task", schema=None) as batch_op:
-        for column, old_type, new_type, nullable in TASK_TIMESTAMP_COLUMN_SPECS:
-            batch_op.drop_column(column)
-            target_type = new_type if is_upgrade else old_type
-            batch_op.alter_column(
-                _rep_col(column),
-                new_column_name=column,
-                existing_type=target_type,
-                existing_nullable=True,
-                nullable=nullable,
-            )
-
-
-def _copy_sqlite_task_timestamps_for_upgrade() -> None:
-    """Copy existing SQLite task timestamp values into upgraded columns."""
-    op.execute(
-        f"""
-        UPDATE task
-        SET {_rep_col("active_until")} = datetime(active_until, 'unixepoch') || '+00:00'
-        WHERE active_until IS NOT NULL
-        """
-    )
-    for column in TASK_TIMESTAMP_COLUMNS:
-        # SQLite TIMESTAMP storage uses a space as the date/time separator.
-        op.execute(
-            f"""
-            UPDATE task
-            SET {_rep_col(column)} = replace({column}, 'T', ' ')
-            WHERE {column} IS NOT NULL
-            """
-        )
-
-
-def _copy_sqlite_task_timestamps_for_downgrade() -> None:
-    """Copy existing SQLite task timestamp values into downgraded columns."""
-    op.execute(
-        f"""
-        UPDATE task
-        SET {_rep_col("active_until")} = CAST(strftime('%s', active_until) AS BIGINT)
-        WHERE active_until IS NOT NULL
-        """
-    )
-    for column in TASK_TIMESTAMP_COLUMNS:
-        # Downgraded string values use datetime.isoformat() style separators.
-        op.execute(
-            f"""
-            UPDATE task
-            SET {_rep_col(column)} = replace({column}, ' ', 'T')
-            WHERE {column} IS NOT NULL
-            """
-        )
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    _add_replacement_columns(is_upgrade=True)
-    if op.get_bind().dialect.name == "sqlite":
-        _copy_sqlite_task_timestamps_for_upgrade()
-    _replace_task_timestamp_columns(is_upgrade=True)
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    _add_replacement_columns(is_upgrade=False)
-    if op.get_bind().dialect.name == "sqlite":
-        _copy_sqlite_task_timestamps_for_downgrade()
-    _replace_task_timestamp_columns(is_upgrade=False)
-
-    # ### end Alembic commands ###
