--- framework/py/flwr/supercore/state/alembic/versions/rev_2026_05_15_drop_run_status_columns.py
@@ -1,74 +1,0 @@
-# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Drop run status columns.
-
-Revision ID: 26b5d6018750
-Revises: e937d3528d23
-Create Date: 2026-05-15 11:22:41.096410
-"""
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# pylint: disable=no-member
-
-# revision identifiers, used by Alembic.
-revision: str = "26b5d6018750"
-down_revision: str | Sequence[str] | None = "e937d3528d23"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.drop_column("pending_at")
-        batch_op.drop_column("starting_at")
-        batch_op.drop_column("running_at")
-        batch_op.drop_column("finished_at")
-        batch_op.drop_column("sub_status")
-        batch_op.drop_column("details")
-
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table("run", schema=None) as batch_op:
-        batch_op.add_column(sa.Column("pending_at", sa.VARCHAR(), nullable=True))
-        batch_op.add_column(sa.Column("starting_at", sa.VARCHAR(), nullable=True))
-        batch_op.add_column(sa.Column("running_at", sa.VARCHAR(), nullable=True))
-        batch_op.add_column(sa.Column("finished_at", sa.VARCHAR(), nullable=True))
-        batch_op.add_column(
-            sa.Column(
-                "sub_status",
-                sa.VARCHAR(),
-                server_default=sa.text("''"),
-                nullable=False,
-            )
-        )
-        batch_op.add_column(
-            sa.Column(
-                "details",
-                sa.VARCHAR(),
-                server_default=sa.text("''"),
-                nullable=False,
-            )
-        )
-
-    # ### end Alembic commands ###
